Added the "function" tag to ApiExtractor.
This change the behaviour of ApiExtractor regarding to global functions.
All global function you want to be exported to python *need* to be especified
in the type system with the function tag, otherwise they wont be exported at all.

The syntax for this new tag is:

<function signature="..." />

This is just the initial work for this tag, it is missign support for:

- Function modifications.
- Add a function overload with add-function tag.