Added test for overload decisor handling container dependencies.
Explaining with an example. Consider a function called "function" with
two signatures accepting a list of Point and a list of PointF,
respectively. Consider also that Point is implicitly convertible to
PointF.

void function(list<Point>&)
void function(list<PointF>&)

A list of Point should be checked before a list of PointF.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>
Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>