Updated the overload sorting to handle convertible types inside containers.
The code as it is now will work for containers with wrapped types,
and will also consider some primitive types implicitly convertible
for dependency sorting purposes.

One weakness of the current solution is that it doesn't handle all
possible combinations for containers with multiple types. E.g. for
Container<T1, T2>, with T1 being implicitly convertible from I1, and T2
from I2, dependency resolution will take into account

* Container<T1, T2>
* Container<I1, T2>
* Container<T1, I2>

but not

* Container<I1, I2>

For the time being every binding is working fine this way, when the need
arises we can fix it, for now let us not slow down the generation.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>
Reviewed by Renato Araújo <renato.filho@openbossa.org>