Append python name to libshiboken output file
libshiboken is strongly tied to the python
it was compiled against. This commit adds
the python name to the output file to make
this connection explicit.

The generator plugin, binary and includes are
untouched as they don't depend on python.

Also, Instead of installing the cmake info in a
single file, ShibokenConfig.cmake will load the
correct file (ShibokenConfig-<python name>.cmake)
based on the value of PYTHON_BASENAME when cmake
is called.

The last shiboken installed will be the default
as each install will overwrite ShibokenConfig.cmake.

To select an specific python, call cmake with
-DPYTHON_BASENAME=python2.6, for python2.6 release.

Reviewer: Marcelo Lira <marcelo.lira@openbossa.org>
          Lauro Moura <lauro.neto@openbossa.org>