Added convenience functions to improve code readability.
Checking if a type is an Object Type is a very common task, followed
by asking if a type is a pointer to a type that has a Python wrapper.

These functions solve the problem:

ShibokenGenerator::isObjectType(type)
ShibokenGenerator::isPointerToWrapperType(type)

I refactored the generator code to make use of those functions.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>
Reviewed by Lauro Moura <lauro.neto@openbossa.org>
Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>