Fix BUG #1086 - "generatorrunner segfault processing #include..."
Boost headers typically include the character '#' at the beginning of
any line and just do nothing elese with that unnamed directive. So as
that's not actually an error, rather that's just a just-do-nothing
directive we'll ignore this situtation. There's still a problem when
using #include directives and passing macros as arguments rather than a
path to the header file. Instead of asserting this issue we'll print a
warning message about this issue.

One can also include a path between brackets (for e.g., when defining a
macro or so) so that we cannot simply ignore that. So we need to handle
this kind of stuff as well.

See http://bugs.pyside.org/show_bug.cgi?id=1086.

Signed-off-by: Paulo Alcantara <pcacjr@gmail.com>
Reviewed-by: Hugo Parente Lima <hugo.pl@gmail.com>
Reviewed-by: Marcelo Lira <marcelo.lira@openbossa.org>