BindingManagerPrivate::assignWrapper will override existing bindings.
This is needed to handle the case where the underlying C++ object is deleted, but its Python wrapper is still alive. If a new C++ object is allocated at the same address, it needs to override the old binding to prevent the situation where the old binding is reused. If the old binding is not replaced, a C++ object may be bound to a Python wrapper for different object type. (Or worse, a C++ object is bound to a Python object which no longer exists.)